requirements.txt
django==1.10
django-registration-redux=1.4